Fix code analysis
Dispose resources via IDisposable
